Serve JSON-encoded server stats on /rcstream_status
Handle ordinary HTTP requests to /rcstream_status by responding with
JSON-encoded server stats. This could easily be restricted to internal clients,
but since the datapoints are not sensitive, I've kept it open.

My intent is to rig up a Diamond collector that works by polling this URL on
each rcstream instance, a bit like the Nginx collector (polls for Nginx stats
served on /nginx_status).

Change-Id: Ia9de677c8af1fd32017fc06e45b4f361c4c445f3